Move one_on_one to separate file
